Api

Description

Provides a RESTful API for GETting all possible representations using the number to letter mapping provided in the Dialpad module.

Making Requests

GET requests must be made to / with query parameter input as an array of integers.

Format of the input parameter

An array containing the numbers 2 and 5 must be formatted as input[]=2&input=5. For Example:

>>> curl -v -X GET localhost:3000/?input[]=2&input[]=5
* Rebuilt URL to: localhost:3000/?input[]=2&input[]=5
*   Trying 127.0.0.1...
* Connected to localhost (127.0.0.1) port 3000 (#0)
> GET /?input=25 HTTP/1.1
> Host: localhost:3000
> User-Agent: curl/7.50.1
> Accept: */*
> 
< HTTP/1.1 200 OK
< Transfer-Encoding: chunked
< Date: Sat, 02 Sep 2017 04:39:50 GMT
< Server: Warp/3.2.13
< Content-Type: application/json;charset=utf-8
< 
* Connection #0 to host localhost left intact
["aj","ak","al","bj","bk","bl","cj","ck","cl"]

Types of Responses

200 OK

Returns the result of telephoneWords in the response body as JSON

400 Bad Request

Occurs when the input parameter is malformed, i.e. not a valid sequence of phone digits as defined by the Dialpad module

Documentation

app :: Application

Used by the executable in main to serve requests