Dialpad

Description

A library for handling telephone digit inputs.

A Valid Sequence of Phone Digits

A list of integers represents a valid sequence of phone digits if it only contains one-digit non-negative integers

The following are all invalid sequences:

[2,15]

[-1, 2, 5]

Documentation

data PhoneDigit

A type representing the digit buttons of a telephone keypad

Constructors

Zero

One

Two

Three

Four

Five

Six

Seven

Eight

Nine

telephoneMapping :: PhoneDigit -> [Char]

Maps a digit to a list of the characters it represents.

This mapping is consistent with that of a modern phone, which maps letters Q and Z to numbers 7 and 9 respectively.

hasMapping :: PhoneDigit -> Bool

Determines whether a phone digit key corresponds to any letters.

digitsFromIntList :: [Int] -> [PhoneDigit]

Convert a string to PhoneDigits if that string represents a valid sequence of phone digits.

>>> digitsFromString "25"
Just [Two, Five]

>>> digitsFromString "ab34"
Nothing

>>> digitsFromString "2,5"
Nothing

isValidSequence :: [Int] -> Bool

Given a list of integers, return True if it represents a valid sequence of phone digits, or False otherwise

telephoneWords :: [Int] -> Maybe [String]

Given a list of ints that represents a valid sequence of phone digits, return a list of all possible words (including the gibberish ones!) that the dialpad sequence can map to

>>> telephoneWords [2,5]
Just ["aj","ak","al","bj","bk","bl","cj","ck","cl"]

>>> telephoneWords [2,5,25]
Nothing

>>> telephoneWords [0]
Just []