Dialpad

Description

A Valid Sequence of Phone Digits

This module currently considers a valid sequence of phone digits to be a string that only contains numeric characters, or is the empty string ""

The following are all invalid sequences:

"a25"

"2,5"

Documentation

data PhoneDigit

A type representing the digit buttons of a telephone keypad

Constructors

Zero

One

Two

Three

Four

Five

Six

Seven

Eight

Nine

digitsFromString :: String -> Maybe [PhoneDigit]

Convert a string to PhoneDigits if that string represents a valid sequence of phone digits.

>>> digitsFromString "25"
Just [Two, Five]

>>> digitsFromString "ab34"
Nothing

>>> digitsFromString "2,5"
Nothing

telephoneWords :: String -> Maybe [String]

Given an input string that represents a sequence of phone digits, return a list of all possible words (including the gibberish ones!) that the dialpad sequence can map to

>>> telephoneWords "25"
Just ["aj","ak","al","bj","bk","bl","cj","ck","cl"]

>>> telephoneWords "ab34"
Nothing

>>> telephoneWords "2,5"
Nothing